requirements.txt
absl-py==0.8.1
appnope==0.1.0
astor==0.8.0
attrs==19.3.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==3.1.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.6
Flask==1.1.1
funcsigs==1.0.2
gast==0.2.2
google
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==19.10.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.2.0
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
mock==3.0.5
more-itertools==5.0.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.5
oauthlib==3.1.0
opt-einsum==2.3.2
pandocfilters==1.4.2
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.11.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scandir==1.10.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
subprocess32==3.5.4
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
